+++ hazelcast/src/main/java/com/hazelcast/internal/monitors/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Contains functionality what is happening in a HazelcastInstance.
+ */
+package com.hazelcast.internal.monitors;
